New file DesignTimeMvcBuilderConfiguration.cs
