+++ src/Orchard.Cms.Web/Startup.cs
@@ -1,10 +1,6 @@
-using System.Collections.Generic;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Builder.Internal;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Modules.Hosting;
-using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -12,7 +8,6 @@ using Orchard.DisplayManagement;
 using Orchard.Environment.Extensions;
 using Orchard.Environment.Extensions.Manifests;
 using Orchard.Environment.Shell.Data;
-using Orchard.Hosting.Mvc.Razor;
 
 namespace Orchard.Cms.Web
 {
@@ -47,36 +42,6 @@ namespace Orchard.Cms.Web
             loggerFactory.AddConsole(Configuration);
 
             app.UseModules();
-        }
-    }
-
-    public class DesignTimeMvcBuilderConfiguration : IDesignTimeMvcBuilderConfiguration
-    {
-        public void ConfigureMvc(IMvcBuilder builder)
-        {
-            var serviceProvider = builder.Services.BuildServiceProvider();
-            var env = serviceProvider.GetRequiredService<IHostingEnvironment>();
-
-            var services = new ServiceCollection();
-            services.AddSingleton<IHostingEnvironment>(env);
-
-            var startUp = new Startup(env);
-            startUp.ConfigureServices(services);
-
-            serviceProvider = services.BuildServiceProvider();
-            var loggerFactory = serviceProvider.GetRequiredService<ILoggerFactory>();
-
-            var app = new ApplicationBuilder(serviceProvider);
-            startUp.Configure(app, loggerFactory);
-
-            builder.AddRazorOptions(options =>
-            {
-                var expander = new ModuleViewLocationExpander();
-                options.ViewLocationExpanders.Add(expander);
-
-                var extensionLibraryService = services.BuildServiceProvider().GetService<IExtensionLibraryService>();
-                ((List<MetadataReference>)options.AdditionalCompilationReferences).AddRange(extensionLibraryService.MetadataReferences());
-            });
         }
     }
 }
